+++ changes.md
@@ -1,5 +1,7 @@
 ### v0.7.*@dev (master)
 
+### v0.7.3
+
 #### Removed
 
 - Default throttles for rate limiting have been removed.
@@ -7,6 +9,7 @@
 #### Fixed
 
 - Fixed bug where any response object that could not be type cast as a string would throw an `UnexpectedValueException`.
+- Fixed bug where request was not bound as an instance resulting in rebounds not being fired.
 
 ### v0.7.2